Fix server killing

+++ index.js
@@ -3,8 +3,9 @@ const clear = require('clear');
 const spawn = require('child_process').spawn;
 const gamedig = require('gamedig');
 
-const logError = (str) => console.log(chalk.red(`(!) ${ str }`));
-const logMsg = (str) => console.log(chalk.blue(`(*) ${ str }`));
+const logError = (str) => console.log(chalk.red(`Watchdog: (!) ${str}`));
+const logMsg = (str) => console.log(chalk.blue(`Watchdog: (*) ${str}`));
+const logChild = (str) => console.log(chalk.yellow(`Child (PID: ${child.pid}): ${str}`));
 
 clear();
 const args = process.argv.slice(2);
@@ -12,7 +13,7 @@ const args = process.argv.slice(2);
 const port = Number.parseInt(args[0]);
 const script = args[1];
 
-if (Number.isNaN(port) || port < 1000 || port > 55565) {
+if (Number.isNaN(port) || port < 1000 || port > 75565) {
     logError('Invalid port');
     process.exit(1);
 } else if (!script) {
@@ -27,11 +28,15 @@ let failed = 0;
 const spawnProcess = async () => {
     // If the child is running, kill it and wait 5 seconds
     if (child) {
-        child.kill('SIGKILL');
+        if (!child.killed) {
+            logMsg('Previous unstopped child process detected (how?)... sending SIGKILL');
+            child.kill('SIGKILL');
 
-        await new Promise(resolve => {
-            setTimeout(() => resolve(), 5000);
-        });
+            await new Promise(resolve => {
+                setTimeout(() => resolve(), 5000);
+            });
+        }
+
         child = undefined;
     }
 
@@ -44,7 +49,7 @@ const spawnProcess = async () => {
 
     child.stdout.setEncoding('utf8');
     child.stdout.on('data', data => {
-        logMsg(data);
+        logChild(data);
 
         if (!running) {
             running = true;
@@ -54,16 +59,20 @@ const spawnProcess = async () => {
     child.stderr.setEncoding('utf8');
     child.stderr.on('data', data => logError(data));
 
-    child.on('close', () => {
-        // logError('Unexpected process termination. Rebooting server!');
-        // spawnProcess();
+    child.on('close', async () => {
+        logError('Process ended. Rebooting server!');
+        await new Promise(resolve => {
+            setTimeout(() => resolve(), 5000);
+        });
+
+        spawnProcess();
     });
 };
 
 spawnProcess();
 
 setInterval(async () => {
-    if (!running) {
+    if (!running || !child) {
         return;
     }
 
@@ -80,7 +89,8 @@ setInterval(async () => {
         if (failed >= 8) {
             logError('Force restarting server');
             failed = 0;
-            await spawnProcess();
+
+            child.kill('SIGKILL'); // kill the process
         } else {
             failed++;
             logError('Failed to ping server! (' + failed + '/8)');
